[fix] intermittent adding to wrong base pane
-invalid 'session only' target syntax in use - targetting is based on
pane 'addresses'! if the target pane is part of another session then
that session can be prefixed, but it's still always only a pane. the
'session-only' syntax was thus actually specifying a pane, and given
the session# will have been different on various attempts to fix this
bug, this was the cause of the 'intermittent' status, sometimes the
'session #' == 'target pane idx', sometimes not